# api/heytom-auth/api/auth/v1/application_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v6.31.1
// source: auth/v1/application.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationApplicationCreate = "/auth.v1.Application/Create"
const OperationApplicationDelete = "/auth.v1.Application/Delete"
const OperationApplicationGet = "/auth.v1.Application/Get"
const OperationApplicationPageList = "/auth.v1.Application/PageList"
const OperationApplicationUpdate = "/auth.v1.Application/Update"

type ApplicationHTTPServer interface {
	Create(context.Context, *CreateApplicationRequest) (*ApplicationInfo, error)
	Delete(context.Context, *DeleteApplicationRequest) (*emptypb.Empty, error)
	Get(context.Context, *GetApplicationRequest) (*ApplicationInfo, error)
	PageList(context.Context, *PageApplicationRequest) (*PageApplicationResponse, error)
	Update(context.Context, *UpdateApplicationRequest) (*emptypb.Empty, error)
}

func RegisterApplicationHTTPServer(s *http.Server, srv ApplicationHTTPServer) {
	r := s.Route("/")
	r.GET("/application/list", _Application_PageList0_HTTP_Handler(srv))
	r.GET("/application/{id}", _Application_Get0_HTTP_Handler(srv))
	r.POST("/application", _Application_Create0_HTTP_Handler(srv))
	r.PUT("/application/{id}", _Application_Update0_HTTP_Handler(srv))
	r.DELETE("/application/{id}", _Application_Delete0_HTTP_Handler(srv))
}

func _Application_PageList0_HTTP_Handler(srv ApplicationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in PageApplicationRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationApplicationPageList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.PageList(ctx, req.(*PageApplicationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*PageApplicationResponse)
		return ctx.Result(200, reply)
	}
}

func _Application_Get0_HTTP_Handler(srv ApplicationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetApplicationRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationApplicationGet)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Get(ctx, req.(*GetApplicationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ApplicationInfo)
		return ctx.Result(200, reply)
	}
}

func _Application_Create0_HTTP_Handler(srv ApplicationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateApplicationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationApplicationCreate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Create(ctx, req.(*CreateApplicationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ApplicationInfo)
		return ctx.Result(200, reply)
	}
}

func _Application_Update0_HTTP_Handler(srv ApplicationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateApplicationRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationApplicationUpdate)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Update(ctx, req.(*UpdateApplicationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _Application_Delete0_HTTP_Handler(srv ApplicationHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteApplicationRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationApplicationDelete)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Delete(ctx, req.(*DeleteApplicationRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type ApplicationHTTPClient interface {
	Create(ctx context.Context, req *CreateApplicationRequest, opts ...http.CallOption) (rsp *ApplicationInfo, err error)
	Delete(ctx context.Context, req *DeleteApplicationRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	Get(ctx context.Context, req *GetApplicationRequest, opts ...http.CallOption) (rsp *ApplicationInfo, err error)
	PageList(ctx context.Context, req *PageApplicationRequest, opts ...http.CallOption) (rsp *PageApplicationResponse, err error)
	Update(ctx context.Context, req *UpdateApplicationRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type ApplicationHTTPClientImpl struct {
	cc *http.Client
}

func NewApplicationHTTPClient(client *http.Client) ApplicationHTTPClient {
	return &ApplicationHTTPClientImpl{client}
}

func (c *ApplicationHTTPClientImpl) Create(ctx context.Context, in *CreateApplicationRequest, opts ...http.CallOption) (*ApplicationInfo, error) {
	var out ApplicationInfo
	pattern := "/application"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationApplicationCreate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ApplicationHTTPClientImpl) Delete(ctx context.Context, in *DeleteApplicationRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/application/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationApplicationDelete))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ApplicationHTTPClientImpl) Get(ctx context.Context, in *GetApplicationRequest, opts ...http.CallOption) (*ApplicationInfo, error) {
	var out ApplicationInfo
	pattern := "/application/{id}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationApplicationGet))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ApplicationHTTPClientImpl) PageList(ctx context.Context, in *PageApplicationRequest, opts ...http.CallOption) (*PageApplicationResponse, error) {
	var out PageApplicationResponse
	pattern := "/application/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationApplicationPageList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *ApplicationHTTPClientImpl) Update(ctx context.Context, in *UpdateApplicationRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/application/{id}"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationApplicationUpdate))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}