# openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /application:
        post:
            tags:
                - Application
            operationId: Application_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.CreateApplicationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.ApplicationInfo'
    /application/list:
        get:
            tags:
                - Application
            operationId: Application_PageList
            parameters:
                - name: page.pageIndex
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: page.pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.PageApplicationResponse'
    /application/{id}:
        get:
            tags:
                - Application
            operationId: Application_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.ApplicationInfo'
        put:
            tags:
                - Application
            operationId: Application_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.UpdateApplicationRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
        delete:
            tags:
                - Application
            operationId: Application_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content: {}
    /auth/github:
        post:
            tags:
                - Auth
            operationId: Auth_LoginByGithub
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.LoginByGithubRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.TokenResponse'
    /auth/login:
        post:
            tags:
                - Auth
            operationId: Auth_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.TokenResponse'
    /auth/logout:
        post:
            tags:
                - Auth
            operationId: Auth_Logout
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /auth/register:
        post:
            tags:
                - Auth
            operationId: Auth_Register
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/auth.v1.RegisterRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/auth.v1.TokenResponse'
...
components:
    schemas:
        auth.v1.ApplicationInfo:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                appKey:
                    type: string
                appSecret:
                    type: string
                callbackUrl:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 应用相关消息
        auth.v1.CreateApplicationRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                callbackUrl:
                    type: string
...
        auth.v1.LoginByGithubRequest:
            type: object
            properties:
                code:
                    type: string
        auth.v1.LoginRequest:
            type: object
            properties:
                mobile:
                    type: string
                password:
                    type: string
...
        auth.v1.PageApplicationResponse:
            type: object
            properties:
                total:
                    type: string
                applications:
                    type: array
                    items:
                        $ref: '#/components/schemas/auth.v1.ApplicationInfo'
...
        auth.v1.RegisterRequest:
            type: object
            properties:
                name:
                    type: string
                mobile:
                    type: string
                password:
                    type: string
...
        auth.v1.TokenResponse:
            type: object
            properties:
                accessToken:
                    type: string
                refreshToken:
                    type: string
                expiresIn:
                    type: string
        auth.v1.UpdateApplicationRequest:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                callbackUrl:
                    type: string
...